Added license name
tl;dr : Added licese name at top of file, as it makes easy to understand which license it is, rather then reading the full text.
---------
Hope I am not being too picky, sending in a one line PR for the license file. Feel free to reject! Personally, When I want to use an opensource project and if I don't see the license name, I tend to read the full license to check if there are any changes or additions to the clauses of standard license text. If I see the license name, I don't read ahead. Mentioning the license name saves time trying to figure out what license it is and if there are any changes to it, specially for noobs.